+++ changelog.md
@@ -1,6 +1,10 @@
 
 ## Changelog for the `groan_rs` library
 
+### Version 0.4.1
+- Added new `TrajMasterRead` trait that is implemented by `TrajReader`, `TrajRangeReader`, `TrajStepReader`, and `TrajRangeStepReader` allowing easier usage of these structures in generic functions.
+- All new traits and structures for trajectory reading are now properly included in the `prelude`.
+
 ### Version 0.4.0
 #### Changes to the Groan selection language
 - **Breaking Change:** Group names can no longer include the characters '<', '>', or '='.